Added polish + blorb color
Added some nice comments for clarity

Gave blorbs a proper jiggleIntensity field for how frisky they are in
the field

Added this.fillClr and this.strokeClr for drawing blorbs with
programmatically adjustable colors, however it seems to be busted at the
moment. Need to troubleshoot.

Removed redundant pellet spawn loop from script body as pelletSpawn
already covers it.